requirements.txt
flask
pymongo[srv]==3.12.0
python-dotenv
flask_cors
flask-marshmallow